split archiving functions from general helpers
